feat: public page layout and design

+++ app/api/posts/[id]/route.ts
@@ -1,5 +1,6 @@
 import { NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
+import { getUserFromRequest } from "@/lib/auth";
 
 export async function GET(
   _req: Request,
@@ -25,10 +26,29 @@ export async function PUT(
   req: Request,
   { params }: { params: Promise<{ id: string }> },
 ) {
+  const user = await getUserFromRequest();
+
+  if (!user) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
   const { id } = await params;
   const body = await req.json();
 
-  const post = await prisma.post.update({
+  const post = await prisma.post.findUnique({
+    where: { id },
+    select: { authorId: true },
+  });
+
+  if (!post) {
+    return NextResponse.json({ error: "Post not found" }, { status: 404 });
+  }
+
+  if (post.authorId !== user.userId) {
+    return NextResponse.json({ message: "Forbidden" }, { status: 403 });
+  }
+
+  const updatedPost = await prisma.post.update({
     where: { id },
     data: {
       title: body.title,
@@ -37,16 +57,33 @@ export async function PUT(
     },
   });
 
-  return NextResponse.json({ data: post });
+  return NextResponse.json({ data: updatedPost });
 }
 
 export async function DELETE(
   _req: Request,
   ctx: { params: Promise<{ id: string }> },
 ) {
+  const user = await getUserFromRequest();
+
+  if (!user) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
   const { id } = await ctx.params;
 
-  console.log("DELETE HIT:", id);
+  const post = await prisma.post.findUnique({
+    where: { id },
+    select: { authorId: true },
+  });
+
+  if (!post) {
+    return NextResponse.json({ error: "Post not found" }, { status: 404 });
+  }
+
+  if (post.authorId !== user.userId) {
+    return NextResponse.json({ message: "Forbidden" }, { status: 403 });
+  }
 
   await prisma.post.delete({
     where: { id },
